use consistent library description phrase
The (Unofficial) looks kinda weird and is pretty well redundant
because this isn't under the elixir-mint organization. Plus what
does it mean to be official anyways 🙃. It may be a bit deceiving
on hex.pm that this isn't from the same authors as Mint, but there's
an ownership section in Hex so I'm not too concerned.